[#12297] test(secret): Split SecretBinding and SecretReference unit tests
