+++ api/src/test/java/org/apache/gravitino/secret/TestSecretReference.java
@@ -23,19 +23,7 @@ import java.util.Map;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
-public class TestSecretBindingAndReference {
-
-  @Test
-  void testSecretBinding() {
-    SecretBinding binding = new SecretBinding("memory", "s3cr3t");
-    Assertions.assertEquals("memory", binding.provider());
-    Assertions.assertEquals("s3cr3t", binding.plaintext());
-    Assertions.assertFalse(binding.toString().contains("s3cr3t"));
-
-    Assertions.assertThrows(
-        IllegalArgumentException.class, () -> new SecretBinding("memory", "******"));
-    Assertions.assertThrows(IllegalArgumentException.class, () -> new SecretBinding("", "x"));
-  }
+public class TestSecretReference {
 
   @Test
   void testSecretReference() {
